add table component
