+++ packages/playground/src/portal/dedicated_nodes_view.vue
@@ -8,23 +8,13 @@
 
       <!-- Filters -->
       <dedicated-node-filters :filters="filters" />
-      <div class="pt-5">
-        <v-card>
-          <v-tabs v-model="activeTab" align-tabs="center">
-            <v-tab :value="0" color="primary" class="pr-8">Rentable</v-tab>
-            <v-tab :value="1" color="primary" class="pr-8">Rented</v-tab>
-            <v-tab :value="2" color="primary" class="pr-8">Mine</v-tab>
-          </v-tabs>
-        </v-card>
-      </div>
+      <!-- Table -->
+      <nodes-table />
     </v-container>
   </div>
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
-const activeTab = ref(0);
-
 const filters = [
   {
     label: "Total SRU (GB)",
@@ -61,11 +51,13 @@ const filters = [
 
 <script lang="ts">
 import DedicatedNodeFilters from "../components/dedicated_node_filter.vue";
+import NodesTable from "../components/nodes_table.vue";
 
 export default {
   name: "Dedicated Node",
   components: {
     DedicatedNodeFilters,
+    NodesTable,
   },
 };
 </script>
